@@ -2,13 +2,14 @@
   "name": "react-pgn-viewer",
   "version": "0.2.4",
   "description": "React component for creating an interactive pgn viewer from string",
+  "source": "src/index.js",
   "main": "dist/index.js",
   "directories": {
     "test": "test"
   },
   "scripts": {
-    "build": "NODE_ENV=production gulp prod",
-    "dev": "NODE_ENV=development gulp dev",
+    "build": "NODE_ENV=production rollup --config rollup.config.js",
+    "dev": "NODE_ENV=development rollup --config rollup.config.js",
     "lint": "eslint ./",
     "lint-fix": "eslint --fix ./"
   },
@@ -26,12 +27,19 @@
   },
   "homepage": "https://github.com/ajGingrich/react-pgn-viewer#readme",
   "dependencies": {
-    "chess.js": "^0.10.2",
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2",
-    "reactjs-chessboard": "0.0.8"
+    "chess.js": "^0.13.2",
+    "react-chessboard": "^1.2.4"
+  },
+  "peerDependencies": {
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0"
   },
   "devDependencies": {
+    "@babel/preset-react": "^7.17.12",
+    "@rollup/plugin-babel": "^5.3.1",
+    "@rollup/plugin-commonjs": "^22.0.0",
+    "@rollup/plugin-node-resolve": "^13.3.0",
+    "@rollup/plugin-replace": "^4.0.0",
     "babel-cli": "^6.26.0",
     "babel-core": "^6.26.0",
     "babel-eslint": "^8.2.3",
@@ -40,19 +48,23 @@
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-es2015-rollup": "^3.0.0",
     "babel-preset-react": "^6.24.1",
-    "del": "^3.0.0",
+    "del": "^6.1.0",
     "eslint": "^4.19.1",
     "eslint-config-google": "^0.9.1",
     "eslint-plugin-react": "^7.8.2",
     "gulp": "^4.0.2",
     "gulp-webserver": "^0.9.1",
     "highlight.js": "^9.12.0",
-    "rollup": "^0.59.1",
+    "rollup": "^2.73.0",
     "rollup-analyzer-plugin": "^1.1.1",
-    "rollup-plugin-babel": "^3.0.4",
+    "rollup-plugin-babel": "^4.4.0",
+    "rollup-plugin-cleaner": "^1.0.0",
     "rollup-plugin-commonjs": "8.3.0",
+    "rollup-plugin-generate-html-template": "^1.7.0",
+    "rollup-plugin-livereload": "^2.0.5",
     "rollup-plugin-node-resolve": "^3.3.0",
     "rollup-plugin-replace": "^2.0.0",
+    "rollup-plugin-serve": "^1.1.0",
     "rollup-plugin-uglify": "^3.0.0",
     "rollup-stream": "^1.24.1",
     "vinyl-source-stream": "^2.0.0"
