@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import Chess from 'chess.js'
-import Chessboard from 'reactjs-chessboard'
+import Chessboard from 'react-chessboard'
 import BoardHeader from './BoardHeader'
 import BoardFooter from './Footer/BoardFooter'
 import MoveList from './Moves/MoveList'
